@@ -6,10 +6,10 @@ class Planta {
         this.produccion = 0,
     });
 
-    String id;
-    String idTest;
-    int estacion;
-    int produccion;
+    String? id;
+    String? idTest;
+    int? estacion;
+    int? produccion;
 
     factory Planta.fromJson(Map<String, dynamic> json) => Planta(
         id: json["id"],
