@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class DesicionesList extends StatelessWidget {
-    const DesicionesList({Key key}) : super(key: key);
+    const DesicionesList({Key? key}) : super(key: key);
 
     
     Future getRegistros() async{
@@ -17,12 +17,12 @@ class DesicionesList extends StatelessWidget {
         return listAcciones;
     }
 
-    Future getDatos(String id) async{
+    Future getDatos(String? id) async{
         
-        Testplaga testplaga= await DBProvider.db.getTestId(id);
+        Testplaga? testplaga= await (DBProvider.db.getTestId(id));
 
-        Finca finca = await DBProvider.db.getFincaId(testplaga.idFinca);
-        Parcela parcela = await DBProvider.db.getParcelaId(testplaga.idLote);
+        Finca? finca = await DBProvider.db.getFincaId(testplaga!.idFinca);
+        Parcela? parcela = await DBProvider.db.getParcelaId(testplaga.idLote);
 
         return [testplaga, finca, parcela];
     }
