@@ -17,12 +17,12 @@ export 'package:app_plaga_enfermedades/src/models/parcela_model.dart';
 
 class DBProvider {
 
-    static Database _database; 
+    static Database? _database; 
     static final DBProvider db = DBProvider._();
 
     DBProvider._();
 
-    Future<Database> get database async {
+    Future<Database?> get database async {
 
         if ( _database != null ) return _database;
 
@@ -134,44 +134,44 @@ class DBProvider {
 
     //ingresar Registros
     nuevoFinca( Finca nuevaFinca ) async {
-        final db  = await database;
-        final res = await db.insert('Finca',  nuevaFinca.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('Finca',  nuevaFinca.toJson() );
         return res;
     }
 
     nuevoParcela( Parcela nuevaParcela ) async {
-        final db  = await database;
-        final res = await db.insert('Parcela',  nuevaParcela.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('Parcela',  nuevaParcela.toJson() );
         return res;
     }
 
     nuevoTestPlaga( Testplaga nuevaPlaga ) async {
-        final db  = await database;
-        final res = await db.insert('TestPlaga',  nuevaPlaga.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('TestPlaga',  nuevaPlaga.toJson() );
         return res;
     }
 
     nuevoPlanta( Planta nuevaPlanta ) async {
-        final db  = await database;
-        final res = await db.insert('Planta',  nuevaPlanta.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('Planta',  nuevaPlanta.toJson() );
         return res;
     }
 
     nuevoExistePlagas( ExistePlaga existePlaga ) async {
-        final db  = await database;
-        final res = await db.insert('ExistePlaga',  existePlaga.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('ExistePlaga',  existePlaga.toJson() );
         return res;
     }
 
     nuevaDecision( Decisiones decisiones ) async {
-        final db  = await database;
-        final res = await db.insert('Decisiones',  decisiones.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('Decisiones',  decisiones.toJson() );
         return res;
     }
 
     nuevaAccion( Acciones acciones ) async {
-        final db  = await database;
-        final res = await db.insert('Acciones',  acciones.toJson() );
+        final db  = await (database);
+        final res = await db!.insert('Acciones',  acciones.toJson() );
         return res;
     }
 
@@ -180,8 +180,8 @@ class DBProvider {
     //Obtener registros
     Future<List<Finca>> getTodasFincas() async {
 
-        final db  = await database;
-        final res = await db.query('Finca');
+        final db  = await (database);
+        final res = await db!.query('Finca');
 
         List<Finca> list = res.isNotEmpty 
                                 ? res.map( (c) => Finca.fromJson(c) ).toList()
@@ -191,8 +191,8 @@ class DBProvider {
 
     Future<List<Parcela>> getTodasParcelas() async {
 
-        final db  = await database;
-        final res = await db.query('Parcela');
+        final db  = await (database);
+        final res = await db!.query('Parcela');
 
         List<Parcela> list = res.isNotEmpty 
                                 ? res.map( (c) => Parcela.fromJson(c) ).toList()
@@ -202,8 +202,8 @@ class DBProvider {
 
     Future<List<Testplaga>> getTodasTestPlaga() async {
 
-        final db  = await database;
-        final res = await db.query('TestPlaga');
+        final db  = await (database);
+        final res = await db!.query('TestPlaga');
 
         List<Testplaga> list = res.isNotEmpty 
                                 ? res.map( (c) => Testplaga.fromJson(c) ).toList()
@@ -213,8 +213,8 @@ class DBProvider {
 
     Future<List<Planta>> getTodasPlantas() async {
 
-        final db  = await database;
-        final res = await db.query('Planta');
+        final db  = await (database);
+        final res = await db!.query('Planta');
 
         List<Planta> list = res.isNotEmpty 
                                 ? res.map( (c) => Planta.fromJson(c) ).toList()
@@ -222,10 +222,10 @@ class DBProvider {
         return list;
     }
 
-    Future<int> countPlanta(String idTest,  int estacion ) async {
+    Future<int?> countPlanta(String idTest,  int estacion ) async {
 
-        final db = await database;
-        int count = Sqflite.firstIntValue(await db.rawQuery("SELECT COUNT(*) FROM Planta WHERE idTest = '$idTest' AND estacion = '$estacion'"));
+        final db = await (database);
+        int? count = Sqflite.firstIntValue(await db!.rawQuery("SELECT COUNT(*) FROM Planta WHERE idTest = '$idTest' AND estacion = '$estacion'"));
         return count;
     
 
@@ -233,8 +233,8 @@ class DBProvider {
 
     Future<List<Decisiones>> getTodasDesiciones() async {
 
-        final db  = await database;
-        final res = await db.query('Decisiones');
+        final db  = await (database);
+        final res = await db!.query('Decisiones');
 
         List<Decisiones> list = res.isNotEmpty 
                                 ? res.map( (c) => Decisiones.fromJson(c) ).toList()
@@ -244,8 +244,8 @@ class DBProvider {
 
     Future<List<Acciones>> getTodasAcciones() async {
 
-        final db  = await database;
-        final res = await db.rawQuery('SELECT DISTINCT idTest FROM Acciones');
+        final db  = await (database);
+        final res = await db!.rawQuery('SELECT DISTINCT idTest FROM Acciones');
 
         List<Acciones> list = res.isNotEmpty 
                                 ? res.map( (c) => Acciones.fromJson(c) ).toList()
@@ -257,28 +257,28 @@ class DBProvider {
     
     
     //REgistros por id
-    Future<Finca> getFincaId(String id) async{
-        final db = await database;
-        final res = await db.query('Finca', where: 'id = ?', whereArgs: [id]);
+    Future<Finca?> getFincaId(String? id) async{
+        final db = await (database);
+        final res = await db!.query('Finca', where: 'id = ?', whereArgs: [id]);
         return res.isNotEmpty ? Finca.fromJson(res.first) : null;
     }
 
-    Future<Parcela> getParcelaId(String id) async{
-        final db = await database;
-        final res = await db.query('Parcela', where: 'id = ?', whereArgs: [id]);
+    Future<Parcela?> getParcelaId(String? id) async{
+        final db = await (database);
+        final res = await db!.query('Parcela', where: 'id = ?', whereArgs: [id]);
         return res.isNotEmpty ? Parcela.fromJson(res.first) : null;
     }
 
-    Future<Testplaga> getTestId(String id) async{
-        final db = await database;
-        final res = await db.query('Testplaga', where: 'id = ?', whereArgs: [id]);
+    Future<Testplaga?> getTestId(String? id) async{
+        final db = await (database);
+        final res = await db!.query('Testplaga', where: 'id = ?', whereArgs: [id]);
         return res.isNotEmpty ? Testplaga.fromJson(res.first) : null;
     }
 
-    Future<List<Parcela>> getTodasParcelasIdFinca(String idFinca) async{
+    Future<List<Parcela>> getTodasParcelasIdFinca(String? idFinca) async{
 
-        final db = await database;
-        final res = await db.query('Parcela', where: 'idFinca = ?', whereArgs: [idFinca]);
+        final db = await (database);
+        final res = await db!.query('Parcela', where: 'idFinca = ?', whereArgs: [idFinca]);
         List<Parcela> list = res.isNotEmpty 
                     ? res.map( (c) => Parcela.fromJson(c) ).toList() 
                     : [];
@@ -286,19 +286,19 @@ class DBProvider {
         return list;            
     }
 
-    Future<List<Planta>> getTodasPlantaIdTest(String idTest) async{
-        final db = await database;
-        final res = await db.query('Planta', where: 'idTest = ?', whereArgs: [idTest]);
+    Future<List<Planta>> getTodasPlantaIdTest(String? idTest) async{
+        final db = await (database);
+        final res = await db!.query('Planta', where: 'idTest = ?', whereArgs: [idTest]);
         List<Planta> list = res.isNotEmpty 
                     ? res.map( (c) => Planta.fromJson(c) ).toList() 
                     : [];
         return list;            
     }
    
-    Future<List<Planta>> getTodasPlantasIdTest(String idTest, int estacion) async{
-        final db = await database;
-        final res = await db.rawQuery("SELECT * FROM Planta WHERE idTest = '$idTest' AND estacion = '$estacion'");
-        //final res = await db.query('Planta', where: 'idTest = ?', whereArgs: [idTest]);
+    Future<List<Planta>> getTodasPlantasIdTest(String? idTest, int? estacion) async{
+        final db = await (database);
+        final res = await db!.rawQuery("SELECT * FROM Planta WHERE idTest = '$idTest' AND estacion = '$estacion'");
+        //final res = await db!.query('Planta', where: 'idTest = ?', whereArgs: [idTest]);
         List<Planta> list = res.isNotEmpty 
                     ? res.map( (c) => Planta.fromJson(c) ).toList() 
                     : [];
@@ -308,8 +308,8 @@ class DBProvider {
 
     Future<List<ExistePlaga>> getTodasPlagasIdPlanta(String idPlanta) async {
 
-        final db  = await database;
-        final res = await db.rawQuery("SELECT * FROM ExistePlaga WHERE idPlanta = '$idPlanta'");
+        final db  = await (database);
+        final res = await db!.rawQuery("SELECT * FROM ExistePlaga WHERE idPlanta = '$idPlanta'");
 
         List<ExistePlaga> list = res.isNotEmpty 
                     ? res.map( (c) => ExistePlaga.fromJson(c) ).toList() 
@@ -318,29 +318,28 @@ class DBProvider {
         return list;
     }
 
-    Future<int> getPlagasIdPlanta(String idPlanta, int idplaga) async {
+    Future<int?> getPlagasIdPlanta(String idPlanta, int idplaga) async {
         
-        final db  = await database;
+        final db  = await (database);
         String query = "SELECT existe FROM ExistePlaga WHERE idPlanta = '$idPlanta' AND idPlaga = '$idplaga'";
-        final  res = await db.rawQuery(query);
-        int value = res.isNotEmpty ? res[0]['existe'] : -1;
-        //print(value);
+        final  res = await db!.rawQuery(query);
+        int? value = (res.isNotEmpty ? res[0]['existe'] : -1) as int?;
 
         return value;
     }
 
-    Future<List<Decisiones>> getDecisionesIdTest(String idTest) async{
-        final db = await database;
-        final res = await db.query('Decisiones', where: 'idTest = ?', whereArgs: [idTest]);
+    Future<List<Decisiones>> getDecisionesIdTest(String? idTest) async{
+        final db = await (database);
+        final res = await db!.query('Decisiones', where: 'idTest = ?', whereArgs: [idTest]);
         List<Decisiones> list = res.isNotEmpty 
                                 ? res.map( (c) => Decisiones.fromJson(c) ).toList()
                                 : [];
         return list;
     }
 
-    Future<List<Acciones>> getAccionesIdTest(String idTest) async{
-        final db = await database;
-        final res = await db.query('Acciones', where: 'idTest = ?', whereArgs: [idTest]);
+    Future<List<Acciones>> getAccionesIdTest(String? idTest) async{
+        final db = await (database);
+        final res = await db!.query('Acciones', where: 'idTest = ?', whereArgs: [idTest]);
         List<Acciones> list = res.isNotEmpty 
                                 ? res.map( (c) => Acciones.fromJson(c) ).toList()
                                 : [];
@@ -351,8 +350,8 @@ class DBProvider {
     //List Select
     Future<List<Map<String, dynamic>>> getSelectFinca() async {
        
-        final db  = await database;
-        final res = await db.rawQuery(
+        final db  = await (database);
+        final res = await db!.rawQuery(
             "SELECT id AS value, nombreFinca AS label FROM Finca"
         );
         List<Map<String, dynamic>> list = res.isNotEmpty ? res : [];
@@ -363,8 +362,8 @@ class DBProvider {
     }
     
     Future<List<Map<String, dynamic>>> getSelectParcelasIdFinca(String idFinca) async{
-        final db = await database;
-        final res = await db.rawQuery(
+        final db = await (database);
+        final res = await db!.rawQuery(
             "SELECT id AS value, nombreLote AS label FROM Parcela WHERE idFinca = '$idFinca'"
         );
         List<Map<String, dynamic>> list = res.isNotEmpty ? res : [];
@@ -377,121 +376,121 @@ class DBProvider {
     // Actualizar Registros
     Future<int> updateFinca( Finca nuevaFinca ) async {
 
-        final db  = await database;
-        final res = await db.update('Finca', nuevaFinca.toJson(), where: 'id = ?', whereArgs: [nuevaFinca.id] );
+        final db  = await (database);
+        final res = await db!.update('Finca', nuevaFinca.toJson(), where: 'id = ?', whereArgs: [nuevaFinca.id] );
         return res;
 
     }
 
     Future<int> updateParcela( Parcela nuevaParcela ) async {
 
-        final db  = await database;
-        final res = await db.update('Parcela', nuevaParcela.toJson(), where: 'id = ?', whereArgs: [nuevaParcela.id] );
+        final db  = await (database);
+        final res = await db!.update('Parcela', nuevaParcela.toJson(), where: 'id = ?', whereArgs: [nuevaParcela.id] );
         return res;
 
     }
 
     Future<int> updateTestPlaga( Testplaga nuevaPlaga ) async {
 
-        final db  = await database;
-        final res = await db.update('TestPlaga', nuevaPlaga.toJson(), where: 'id = ?', whereArgs: [nuevaPlaga.id] );
+        final db  = await (database);
+        final res = await db!.update('TestPlaga', nuevaPlaga.toJson(), where: 'id = ?', whereArgs: [nuevaPlaga.id] );
         return res;
 
     }
 
 
     //Conteos analisis
-    Future<double> countPlagaEstacion( String idTest, int estacion, int idPlaga) async {
+    Future<double> countPlagaEstacion( String? idTest, int estacion, int idPlaga) async {
 
-        final db = await database;
+        final db = await (database);
         String query =  "SELECT COUNT(*) FROM TestPlaga "+
                         "INNER JOIN Planta ON TestPlaga.id = Planta.idTest " +
                         "INNER JOIN ExistePlaga ON  Planta.id = ExistePlaga.idPlanta " +
                         "WHERE idTest = '$idTest' AND estacion = '$estacion' AND idPlaga = '$idPlaga' AND existe = 1";
-        int res = Sqflite.firstIntValue(await db.rawQuery(query));
-        double value = res/10;
+        int? res = Sqflite.firstIntValue(await db!.rawQuery(query));
+        double value = res!/10;
         return value;
 
     }
 
-    Future<double> countPlagaTotal( String idTest, int idPlaga) async {
+    Future<double> countPlagaTotal( String? idTest, int idPlaga) async {
 
-        final db = await database;
+        final db = await (database);
         String query =  "SELECT COUNT(*) FROM TestPlaga "+
                         "INNER JOIN Planta ON TestPlaga.id = Planta.idTest " +
                         "INNER JOIN ExistePlaga ON  Planta.id = ExistePlaga.idPlanta " +
                         "WHERE idTest = '$idTest' AND idPlaga = '$idPlaga' AND existe = 1";
-        int res = Sqflite.firstIntValue(await db.rawQuery(query));
-        double value = res/30;
+        int? res = Sqflite.firstIntValue(await db!.rawQuery(query));
+        double value = res!/30;
         return value;
 
     }
 
     Future<double> countDeficiencia( String idTest, int estacion) async {
 
-        final db = await database;
+        final db = await (database);
         String query =  "SELECT COUNT(*) FROM Planta WHERE idTest = '$idTest' AND estacion = '$estacion' AND deficiencia = 1";
-        int res = Sqflite.firstIntValue(await db.rawQuery(query));
-        double value = res/10;
+        int? res = Sqflite.firstIntValue(await db!.rawQuery(query));
+        double value = res!/10;
         return value;
 
     }
 
     Future<double> countTotalDeficiencia( String idTest ) async {
 
-        final db = await database;
+        final db = await (database);
         String query =  "SELECT COUNT(*) FROM Planta WHERE idTest = '$idTest' AND deficiencia = 1";
-        int res = Sqflite.firstIntValue(await db.rawQuery(query));
-        double value = res/30;
+        int? res = Sqflite.firstIntValue(await db!.rawQuery(query));
+        double value = res!/30;
         return value;
 
     }
 
-    Future<double> countProduccion( String idTest, int estacion, int estado) async {
+    Future<double> countProduccion( String? idTest, int estacion, int estado) async {
 
-        final db = await database;
+        final db = await (database);
         String query =  "SELECT COUNT(*) FROM Planta WHERE idTest = '$idTest' AND estacion = '$estacion' AND produccion = '$estado'";
-        int res = Sqflite.firstIntValue(await db.rawQuery(query));
-        double value = res/10;
+        int? res = Sqflite.firstIntValue(await db!.rawQuery(query));
+        double value = res!/10;
         return value;
 
     }
 
-    Future<double> countTotalProduccion( String idTest, int estado ) async {
+    Future<double> countTotalProduccion( String? idTest, int estado ) async {
 
-        final db = await database;
+        final db = await (database);
         String query =  "SELECT COUNT(*) FROM Planta WHERE idTest = '$idTest' AND produccion = '$estado'";
-        int res = Sqflite.firstIntValue(await db.rawQuery(query));
-        double value = res/30;
+        int? res = Sqflite.firstIntValue(await db!.rawQuery(query));
+        double value = res!/30;
         return value;
 
     }
 
     // Eliminar registros
-    Future<int> deleteFinca( String idFinca ) async {
+    Future<int> deleteFinca( String? idFinca ) async {
 
-        final db  = await database;
-        final res = await db.delete('Finca', where: 'id = ?', whereArgs: [idFinca]);
+        final db  = await (database);
+        final res = await db!.delete('Finca', where: 'id = ?', whereArgs: [idFinca]);
         return res;
     }
-    Future<int> deleteParcela( String idParcela ) async {
+    Future<int> deleteParcela( String? idParcela ) async {
 
-        final db  = await database;
-        final res = await db.delete('Parcela', where: 'id = ?', whereArgs: [idParcela]);
-        return res;
-    }
-
-    Future<int> deleteTestPlaga( String idTest ) async {
-
-        final db  = await database;
-        final res = await db.delete('TestPlaga', where: 'id = ?', whereArgs: [idTest]);
+        final db  = await (database);
+        final res = await db!.delete('Parcela', where: 'id = ?', whereArgs: [idParcela]);
         return res;
     }
 
-    Future<int> deletePlanta( String idPlanta ) async {
+    Future<int> deleteTestPlaga( String? idTest ) async {
 
-        final db  = await database;
-        final res = await db.delete('Planta', where: 'id = ?', whereArgs: [idPlanta]);
+        final db  = await (database);
+        final res = await db!.delete('TestPlaga', where: 'id = ?', whereArgs: [idTest]);
+        return res;
+    }
+
+    Future<int> deletePlanta( String? idPlanta ) async {
+
+        final db  = await (database);
+        final res = await db!.delete('Planta', where: 'id = ?', whereArgs: [idPlanta]);
         return res;
     }
 
