@@ -20,7 +20,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
 
     bool _guardando = false;
     int variableVacias = 0;
-    int countPlanta = 0;
+    int? countPlanta = 0;
     var uuid = Uuid();
 
     Planta planta = Planta();
@@ -49,7 +49,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
     @override
     Widget build(BuildContext context) {
 
-        List data = ModalRoute.of(context).settings.arguments;
+        List data = ModalRoute.of(context)!.settings.arguments as List<dynamic>;
         
         planta.idTest = data[1];
         planta.estacion = data[0] ;
@@ -73,16 +73,16 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                                     child: Row(
                                         mainAxisAlignment: MainAxisAlignment.end,
                                         children: [
-                                            Expanded(child: Text('Lista Plagas', style: Theme.of(context).textTheme.headline6
+                                            Expanded(child: Text('Lista Plagas', style: Theme.of(context).textTheme.headline6!
                                                             .copyWith(fontSize: 16, fontWeight: FontWeight.w600))),
                                             Container(
                                                 width: 50,
-                                                child: Text('Si', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                                                child: Text('Si', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                                                                 .copyWith(fontSize: 16, fontWeight: FontWeight.w600,) ),
                                             ),
                                             Container(
                                                 width: 50,
-                                                child: Text('No', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                                                child: Text('No', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                                                                 .copyWith(fontSize: 16, fontWeight: FontWeight.w600)),
                                             ),
                                         ],
@@ -97,19 +97,19 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                                         Expanded(child: Container(),),
                                         Container(
                                             width: 50,
-                                            child: Text('Alta', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                                            child: Text('Alta', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                                             .copyWith(fontSize: 16, fontWeight: FontWeight.w600))
                                             //color: Colors.deepPurple,
                                         ),
                                         Container(
                                             width: 50,
-                                            child: Text('Media', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                                            child: Text('Media', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                                             .copyWith(fontSize: 16, fontWeight: FontWeight.w600))
                                             //color: Colors.deepPurple,
                                         ),
                                         Container(
                                             width: 50,
-                                            child: Text('Baja', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6
+                                            child: Text('Baja', textAlign: TextAlign.center, style: Theme.of(context).textTheme.headline6!
                                             .copyWith(fontSize: 16, fontWeight: FontWeight.w600))
                                         ),
                                     ],
@@ -137,7 +137,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
             
             itemBuilder: (BuildContext context, int index) {
                 
-                String labelPlaga = itemPlagas.firstWhere((e) => e['value'] == '$index', orElse: () => {"value": "1","label": "No data"})['label'];
+                String? labelPlaga = itemPlagas.firstWhere((e) => e['value'] == '$index', orElse: () => {"value": "1","label": "No data"})['label'];
                 int idPlaga = int.parse(itemPlagas.firstWhere((e) => e['value'] == '$index', orElse: () => {"value": "100","label": "No data"})['value']);
                 
                 
@@ -147,13 +147,13 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                         Row(
                             mainAxisAlignment: MainAxisAlignment.end,
                             children: <Widget>[
-                                Expanded(child: Text('$labelPlaga', style: Theme.of(context).textTheme.headline6.copyWith(fontSize: 16, fontWeight: FontWeight.w600))),
+                                Expanded(child: Text('$labelPlaga', style: Theme.of(context).textTheme.headline6!.copyWith(fontSize: 16, fontWeight: FontWeight.w600))),
                                 Transform.scale(
                                     scale: 1.2,
                                     child: Radio(
                                         value: '1',
                                         groupValue: radios[itemPlagas[idPlaga]['value']],
-                                        onChanged: (value){
+                                        onChanged: (dynamic value){
                                             setState(() {
                                                 radios[itemPlagas[idPlaga]['value']] = value;
                                             });
@@ -166,7 +166,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                                     child: Radio(
                                         value:'2',
                                         groupValue: radios[itemPlagas[idPlaga]['value']],
-                                        onChanged: (value){
+                                        onChanged: (dynamic value){
                                             setState(() {
                                                 radios[itemPlagas[idPlaga]['value']] = value;
                                             });
@@ -200,7 +200,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                     child: Radio(
                         value: 1,
                         groupValue: planta.produccion,
-                        onChanged: (value) {
+                        onChanged: (dynamic value) {
                             setState(() {
                                 planta.produccion = value;
                             });
@@ -213,7 +213,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                     child: Radio(
                         value: 2,
                         groupValue: planta.produccion,
-                        onChanged: (value) {
+                        onChanged: (dynamic value) {
                             setState(() {
                                 planta.produccion = value;
                             });
@@ -226,7 +226,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
                         child: Radio(
                         value: 3,
                         groupValue: planta.produccion,
-                        onChanged: (value) {
+                        onChanged: (dynamic value) {
                             setState(() {
                                 planta.produccion = value;
                             });
@@ -248,7 +248,7 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
             
             label: Text('Guardar',
                 style: Theme.of(context).textTheme
-                    .headline6
+                    .headline6!
                     .copyWith(fontWeight: FontWeight.w600, color: Colors.white)
             ),
             padding:EdgeInsets.symmetric(vertical: 13, horizontal: 50),
@@ -316,12 +316,12 @@ class _AgregarPlantaState extends State<AgregarPlanta> {
     void mostrarSnackbar(int variableVacias){
         final snackbar = SnackBar(
             content: Text('Hay $variableVacias Campos Vacios, Favor llene todo los campos',
-                style: Theme.of(context).textTheme.headline6.copyWith(color: Colors.white),
+                style: Theme.of(context).textTheme.headline6!.copyWith(color: Colors.white),
             ),
             duration: Duration(seconds: 2),
         );
         setState(() {_guardando = false;});
-        scaffoldKey.currentState.showSnackBar(snackbar);
+        scaffoldKey.currentState!.showSnackBar(snackbar);
     }
 
 
